working auth and task management

+++ apps/task-management/src/tasks/tasks.service.ts
@@ -4,7 +4,7 @@ import { FilterTaskDto } from './dto/filter-task.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
-import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,8 +14,8 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    getTasks(filterQuery: FilterTaskDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterQuery);
+    getTasks(filterQuery: FilterTaskDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterQuery, user);
     }
 
     // filterTasks(filterQuery: FilterTaskDto) {
@@ -33,8 +33,8 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task with ID ${id} not found.`);
@@ -53,8 +53,8 @@ export class TasksService {
     //     return found;
     // }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
     // createTask(createTaskDto: CreateTaskDto): TaskModel {
@@ -67,8 +67,8 @@ export class TasksService {
     //     return task;
     // }
 
-    async deleteTaskById(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
         if (!result.affected) {
             throw new NotFoundException(`Task with ID ${id} not found.`);
         }
@@ -84,8 +84,8 @@ export class TasksService {
     //     return task;
     // }
 
-    async updateProperty(id: number, item: Partial<Task>): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateProperty(id: number, item: Partial<Task>, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         Object.keys(item).forEach(k => task[k] = item[k]);
         await task.save();
         return task;
